refactor(transloco): :recycle: put transloco functionality in the same folder, refactor inline-loader, extract state
Refs #14
