+++ src/app/transloco/is-lang.function.ts
@@ -1,0 +1,6 @@
+import { langList } from "./lang.state";
+import { Lang } from "./lang.types";
+
+export function isLang(key: unknown): boolean {
+  return !!key && typeof key === 'string' && langList.includes(key as Lang);
+}
